+++ src/modules/datastore/caches_controller.erl
@@ -144,7 +144,8 @@ clear_cache(_MemUsage, TargetMemUse, StoreType, [TimeWindow | Windows]) ->
 %%--------------------------------------------------------------------
 -spec get_hooks_config(Models :: list()) -> list().
 get_hooks_config(Models) ->
-  Methods = [save, get, exists, delete, update, create, fetch_link, add_links, delete_links],
+  Methods = [save, get, exists, delete, update, create, create_or_update,
+    fetch_link, add_links, create_link, delete_links],
   lists:foldl(fun(Model, Ans) ->
     ModelConfig = lists:map(fun(Method) ->
       {Model, Method}